fix(gpu): implement direct constraint propagation on the GPU
DefaultCoordinator::coordinate_propagation was a no-op, so collapsing a cell
never constrained its neighbours and generated grids ignored the adjacency
rules. Wire it to the propagator's strategy and make
DirectPropagationStrategy work:

- bind all ten buffers propagate.wgsl declares, in its order
- make the Params structs in propagate.wgsl and collapse_cell.wgsl mirror
  GpuParamsUniform
- pass the input worklist size through params and reset the output counter
  before every pass
- finish with a full-grid pass that changes nothing, because neighbour
  updates in the shader are not atomic and can lose a restriction
- report where a contradiction happened

Also stop upload_rules from writing rule weights into the adjacency buffer,
stop update_params_worklist_size from zeroing every other parameter, and
match the worklist count binding size to its 4-byte buffer.

Refs #3

+++ wfc-gpu/src/shader/pipeline.rs
@@ -403,7 +403,7 @@ impl ComputePipelines {
                             ty: wgpu::BufferBindingType::Storage { read_only: false },
                             has_dynamic_offset: false,
                             min_binding_size: wgpu::BufferSize::new(
-                                std::mem::size_of::<u64>() as u64
+                                std::mem::size_of::<u32>() as u64
                             ), // Two atomic<u32>
                         },
                         count: None,
